+++ shiny/api-examples/input_text/app-core.py
@@ -2,12 +2,12 @@ from shiny import App, Inputs, Outputs, Session, render, ui
 
 app_ui = ui.page_fluid(
     ui.input_text("caption", "Caption:", "Data summary"),
-    ui.output_text_verbatim("value"),
+    ui.output_code("value"),
 )
 
 
 def server(input: Inputs, output: Outputs, session: Session):
-    @render.text
+    @render.code
     def value():
         return input.caption()