+++ shiny/api-examples/sidebar/app-core-dynamic.py
@@ -58,10 +58,10 @@ def server(input: Inputs, output: Outputs, session: Session):
                 position=input.position(),
             ),
             ui.h2("Dynamic sidebar"),
-            ui.output_text_verbatim("state_dynamic"),
+            ui.output_code("state_dynamic"),
         )
 
-    @render.text
+    @render.code
     def state_dynamic():
         return f"input.sidebar_dynamic(): {input.sidebar_dynamic()}"