+++ shiny/api-examples/theme/app-core-local.py
@@ -10,7 +10,7 @@ app_ui = ui.page_sidebar(
         title="Parameters",
     ),
     ui.h2("Output"),
-    ui.output_text_verbatim("txt"),
+    ui.output_code("txt"),
     ui.markdown(filler_text),
     title="Theme Example",
     theme=Path(__file__).parent / "css" / "shiny-theme-demo.css",
@@ -18,7 +18,7 @@ app_ui = ui.page_sidebar(
 
 
 def server(input, output, session):
-    @render.text
+    @render.code
     def txt():
         return f"n*2 is {input.n() * 2}"