+++ shiny/api-examples/theme/app-core-remote.py
@@ -8,7 +8,7 @@ app_ui = ui.page_sidebar(
         title="Parameters",
     ),
     ui.h2("Output"),
-    ui.output_text_verbatim("txt"),
+    ui.output_code("txt"),
     ui.markdown(filler_text),
     title="Theme Example",
     theme="https://cdn.jsdelivr.net/npm/bootswatch@5.3.3/dist/sketchy/bootstrap.min.css",
@@ -16,7 +16,7 @@ app_ui = ui.page_sidebar(
 
 
 def server(input, output, session):
-    @render.text
+    @render.code
     def txt():
         return f"n*2 is {input.n() * 2}"